--- app/scripts/components/common/layout-root/useFeedbackModal.ts
@@ -1,22 +1,0 @@
-import { useContext, useCallback } from "react";
-import { LayoutRootContext } from "./context";
-
-/**
- * Hook to access the feedback modal.
- */
-export function useFeedbackModal() {
-  const { feedbackModalRevealed, setFeedbackModalRevealed } =
-    useContext(LayoutRootContext);
-
-  return {
-    isRevealed: feedbackModalRevealed,
-    show: useCallback(
-      () => setFeedbackModalRevealed(true),
-      [setFeedbackModalRevealed]
-    ),
-    hide: useCallback(
-      () => setFeedbackModalRevealed(false),
-      [setFeedbackModalRevealed]
-    )
-  };
-}
